Exceptions

+++ back-end/track/src/main/java/com/petfi/track/services/ClientService.java
@@ -3,11 +3,15 @@ package com.petfi.track.services;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.petfi.track.entities.Client;
 import com.petfi.track.repositories.ClientRepository;
+import com.petfi.track.services.exceptions.DatabaseException;
+import com.petfi.track.services.exceptions.ResourceNotFoundException;
 
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 
 @Service
@@ -21,7 +25,7 @@ public class ClientService {
   }
 
   public Client findById(Long id) {
-    return clientRepository.findById(id).get();
+    return clientRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(id));
   }
 
   public Client insert(Client entity) {
@@ -30,9 +34,13 @@ public class ClientService {
 
   @Transactional
   public Client update(Long id, Client entity) {
-    Client client = clientRepository.getReferenceById(id);
-    updateEntity(client, entity);
-    return clientRepository.save(client);
+    try {
+      Client client = clientRepository.getReferenceById(id);
+      updateEntity(client, entity);
+      return clientRepository.save(client);
+    } catch (EntityNotFoundException e) {
+      throw new ResourceNotFoundException(id);
+    }
   }
 
   private void updateEntity(Client client, Client entity) {
@@ -44,6 +52,13 @@ public class ClientService {
   }
 
   public void deleteById(Long id) {
-    clientRepository.deleteById(id);
+    if (!clientRepository.existsById(id)) {
+      throw new ResourceNotFoundException(id);
+    }
+    try {
+      clientRepository.deleteById(id);
+    } catch (DataIntegrityViolationException e) {
+      throw new DatabaseException(e.getMessage());
+    }
   }
 }
